Fixing routes after adding docs path
All the routes (url.py) got messed up by the fact that redirections can
occur with or without a trailing slash. In particular, the socket was
not connecting properly anymore (fixed in sockets.js) and the upload
path was wrong (fixed by making sure that we have trailing slashes
everywhere).

Fixes #88